pyproject.toml
# Build system configuration
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cli_weather"]

# Project metadata and dependencies
[project]
name = "cli-weather"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "geopy==2.4.1",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "tzdata==2024.2",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

# Development dependencies
[project.optional-dependencies]
dev = [
    "ruff>=0.6",        # Linting and formatting
]

# CLI entry point
[project.scripts]
cli-weather = "cli_weather.__main__:main"


# Code formatting and linting
[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]  # Error, pyflakes, warning, import sorting
ignore = ["E501"]              # Ignore line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
